set SYSTEM_TYPE and SYSTEM_FAMILY for zaius
Both of these are turned into the compatible strings for the device-tree
root node and are used by skiboot (and Linux) to identify the platform
they are running on. As a result they need to be well chosen.

Additionally, the "ibm,p9" family string is intended for
IBM FSP based systems so it shouldn't be used here.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>